+++ tests/testthat/test-sfc.R
@@ -126,4 +126,9 @@ test_that("bounding box is flipped when geometry is flipped", {
 	expect_equal(st_bbox(bar), st_bbox(c(xmin=0, ymin=-200, xmax=100, ymax=0)))
 })
 
-
+test_that("st_combine preserves precision of input object", {
+	# See https://github.com/r-spatial/sf/issues/2618
+	pt1 <- st_sfc(st_point(c(1, 1)), st_point(c(2, 2)), precision = 1)
+	pt2 <- st_combine(pt1)
+	expect_equal(st_precision(pt1), st_precision(pt2))
+})